+++ src/level/map/mod.rs
@@ -1,64 +1,112 @@
 // Copyright 2025 Gabriel Bjørnager Jensen.
 
-use crate::level::{
-	Block,
-	Columns,
-	ColumnsMut,
-	MapIter,
-	MapIterMut,
-};
+use crate::level::{Block, Columns, ColumnsMut, MapSize};
 
+use std::mem::swap;
 use std::ops::{Index, IndexMut};
+use std::slice;
 
-#[derive(Debug)]
+#[derive(Clone, Debug, Default)]
 pub struct Map {
-	pub(super) data: Box<[Box<[Block]>]>,
+	height: u32,
+	data:   Box<[Block]>,
 }
 
 impl Map {
 	#[inline]
 	#[must_use]
 	#[track_caller]
-	pub fn new(width: u32, height: u32) -> Self {
-		let columns = vec![Default::default(); height as usize].into_boxed_slice();
+	pub fn new(size: MapSize) -> Self {
+		let data_len = size.product() as usize;
 
-		let map = vec![columns; width as usize].into();
+		let data = vec![Default::default(); data_len].into();
 
-		Self { data: map }
+		Self { height: size.height(), data }
 	}
 
+	#[inline]
+	#[track_caller]
+	pub fn resize(&mut self, size: MapSize) {
+		let data_len = size.product() as usize;
+
+		let mut data = Default::default();
+		swap(&mut data, &mut self.data);
+
+		let mut data: Vec<Block> = data.into();
+
+ 		data.clear();
+		data.resize(data_len, Default::default());
+
+		let mut data = data.into();
+
+		swap(&mut data, &mut self.data);
+
+
+	}
+
+	#[allow(unused)]
 	#[inline(always)]
 	#[must_use]
 	pub fn get(&self, x: u32, y: u32) -> Option<&Block> {
-		let cell = self.data.get(x as usize)?.get(y as usize)?;
-		Some(cell)
+		let index = x as usize * self.height as usize + y as usize;
+		self.data.get(index)
 	}
 
+	#[allow(unused)]
 	#[inline(always)]
 	#[must_use]
 	pub fn get_mut(&mut self, x: u32, y: u32) -> Option<&mut Block> {
-		let cell = self.data.get_mut(x as usize)?.get_mut(y as usize)?;
-		Some(cell)
+		let index = x as usize * self.height as usize + y as usize;
+		self.data.get_mut(index)
 	}
 
+	#[allow(unused)]
 	#[inline(always)]
-	pub fn iter(&self) -> MapIter {
-		MapIter::new(self)
+	pub fn iter(&self) -> slice::Iter<Block> {
+		self.data.iter()
 	}
 
+	#[allow(unused)]
 	#[inline(always)]
-	pub fn iter_mut(&mut self) -> MapIterMut {
-		MapIterMut::new(self)
+	pub fn iter_mut(&mut self) -> slice::IterMut<Block> {
+		self.data.iter_mut()
 	}
 
+	#[allow(unused)]
 	#[inline(always)]
 	pub fn columns(&self) -> Columns {
 		Columns::new(self)
 	}
 
+	#[allow(unused)]
 	#[inline(always)]
 	pub fn columns_mut(&mut self) -> ColumnsMut {
 		ColumnsMut::new(self)
+	}
+
+	#[inline(always)]
+	#[must_use]
+	pub fn height(&self) -> u32 {
+		self.height
+	}
+
+	#[inline(always)]
+	#[must_use]
+	pub fn width(&self) -> u32 {
+		let height = self.height() as usize;
+		(self.data.len() / height) as u32
+	}
+
+	#[inline(always)]
+	#[must_use]
+	pub fn as_ptr(&self) -> *const Block {
+		self.data.as_ptr()
+	}
+
+	#[inline(always)]
+	#[must_use]
+	pub fn as_mut_ptr(&mut self) -> *mut Block {
+		self.data.as_mut_ptr()
 	}
 }
 
@@ -68,7 +116,7 @@ impl Index<(u32, u32)> for Map {
 	#[inline(always)]
 	#[track_caller]
 	fn index(&self, index: (u32, u32)) -> &Self::Output {
-		&self.data[index.0 as usize][index.1 as usize]
+		self.get(index.0, index.1).unwrap()
 	}
 }
 
@@ -76,14 +124,14 @@ impl IndexMut<(u32, u32)> for Map {
 	#[inline(always)]
 	#[track_caller]
 	fn index_mut(&mut self, index: (u32, u32)) -> &mut Self::Output {
-		&mut self.data[index.0 as usize][index.1 as usize]
+		self.get_mut(index.0, index.1).unwrap()
 	}
 }
 
 impl<'a> IntoIterator for &'a Map {
-	type Item = (u32, u32, &'a Block);
+	type Item = &'a Block;
 
-	type IntoIter = MapIter<'a>;
+	type IntoIter = slice::Iter<'a, Block>;
 
 	#[inline(always)]
 	fn into_iter(self) -> Self::IntoIter {
@@ -92,9 +140,9 @@ impl<'a> IntoIterator for &'a Map {
 }
 
 impl<'a> IntoIterator for &'a mut Map {
-	type Item = (u32, u32, &'a mut Block);
+	type Item = &'a mut Block;
 
-	type IntoIter = MapIterMut<'a>;
+	type IntoIter = slice::IterMut<'a, Block>;
 
 	#[inline(always)]
 	fn into_iter(self) -> Self::IntoIter {