+++ src/main.rs
@@ -3,13 +3,31 @@
 #![feature(thread_sleep_until)]
 
 mod app;
+mod error;
 mod level;
+
+use crate::app::App;
+use crate::error::Result;
+
+use std::process::exit;
 
 const _: () = assert!(usize::BITS >= u32::BITS);
 
-fn main() {
-	let config = Default::default();
+fn main() -> ! {
+	let run = || -> Result<()> {
+		let app = App::new()?;
+		app.run()?;
 
-	let app = crate::app::App::new(config);
-	app.run();
+		Ok(())
+	};
+
+	let code = if let Err(e) = run() {
+		eprintln!("error: {e}");
+
+		e.into()
+	} else {
+		0x0
+	};
+
+	exit(code);
 }